Refactor the BreakCommand as an NPC command.
