+++ src/main/java/net/aufdemrand/denizen/scripts/commands/BukkitCommandRegistry.java
@@ -426,39 +426,38 @@ public class BukkitCommandRegistry extends CommandRegistry {
 
         // <--[command]
         // @Name Break
-        // @Syntax break [<location>] (entity:<entity>) (radius:<#.#>)
+        // @Syntax break [<location>] (<npc>) (radius:<#.#>)
         // @Required 1
         // @Stable unstable
-        // @Short Makes the NPC walk over and break a block.
+        // @Short Makes an NPC walk over and break a block.
         // @Author aufdemrand
         // @Group world
         // @Description
         // By itself, the 'break' command will act as a NPC command in the sense that an attached
         // NPC will navigate to and break the block at the attached location. It can also accept a specified npc,
         // to fulfill the command, just specify a 'fetchable' npc object. It can also accept a radius to start
-        // breaking the block from within. To specify the npc, prefix the npc with 'entity:'. To specify the
-        // radius, prefix the radius with 'radius:'.
+        // breaking the block from within. To specify the radius, prefix the radius with 'radius:'.
 
         // @Tags
         // <n@npc.navigator.is_navigating>
         // <n@npc.navigator.target_location>
 
         // @Usage
-        // Use to make a npc break a block at 17,64,-87 in world
+        // Use to make the npc break a block at 17,64,-87 in world.
         // - break l@17,64,-87,world
 
         // @Usage
-        // Use to make a npc with the id 12 break a block at 17,64,-87 in world
-        // - break l@17,64,-87,world entity:n@12
+        // Use to make an npc with the id 12 break a block at 17,64,-87 in world.
+        // - break l@17,64,-87,world n@12
 
         // @Usage
-        // Use to make a npc with the name bob break a block at 17,64,-87 and start digging from 5 blocks away
-        // - break l@17,64,-87,world entity:n@bob radius:5
+        // Use to make an npc with the name bob break a block at 17,64,-87 and start digging from 5 blocks away.
+        // - break l@17,64,-87,world n@bob radius:5
 
         // -->
         if (Depends.citizens != null) {
             registerCoreMember(BreakCommand.class,
-                    "BREAK", "break [<location>] (entity:<entity>) (radius:<#.#>)", 1);
+                    "BREAK", "break [<location>] (<npc>) (radius:<#.#>)", 1);
         }
